+++ pages/about.html.tsx
@@ -1,5 +1,4 @@
 import Head from 'next/head';
-import { useEffect, useState } from 'react';
 import Layout from '@/components/Layout';
 
 
@@ -8,10 +7,6 @@ import Layout from '@/components/Layout';
 import dynamic from 'next/dynamic';
 const MyButton = dynamic(() => import('@/components/Buttons'), { ssr: false });
 
-
-// store
-import { useDispatch, useSelector } from "react-redux";
-import getMenuData from "@/store/actions/demoMenuActions";
 
 
 const MainContent = () => {
@@ -31,31 +26,6 @@ const MainContent = () => {
 */
 const About = () => {
 
-    // Get store
-    const [dispatchUpdate, setDispatchUpdate] = useState<boolean>(false);
-    const dispatch = useDispatch();
-    const storeData = useSelector((state: any) => {
-        return state.menuData;
-    });
-    
-
-
-    useEffect(() => {
-
-        // Get store
-        //-----
-        const fetchStoreMenu = async () => {
-            if ( !dispatchUpdate ) {
-                const res: any = await getMenuData();
-                setDispatchUpdate(true);
-                dispatch(res);
-            }
-        };
-
-        fetchStoreMenu();
-        
-    }, [dispatchUpdate, dispatch]); 
-
     return (
         <>
             <Head>
@@ -65,7 +35,6 @@ const About = () => {
 
             <Layout
                 pageTitle="About"
-                nav={JSON.stringify(storeData.menuItems)}
                 contentComponent={<><MainContent /></>}
             />