+++ pages/nested-routes.html.tsx
@@ -1,12 +1,6 @@
 import Head from 'next/head';
-import { useEffect, useState } from 'react';
 import Link from 'next/link';
 import Layout from '@/components/Layout';
-
-
-// store
-import { useDispatch, useSelector } from "react-redux";
-import getMenuData from "@/store/actions/demoMenuActions";
 
 
 const MainContent = () => {
@@ -36,30 +30,6 @@ const MainContent = () => {
 */
 const NestedRoutes = () => {
 
-    // Get store
-    const [dispatchUpdate, setDispatchUpdate] = useState<boolean>(false);
-    const dispatch = useDispatch();
-    const storeData = useSelector((state: any) => {
-        return state.menuData;
-    });
-    
-
-
-
-    useEffect(() => {
-
-        // Get store
-        //-----
-        const fetchStoreMenu = async () => {
-            if ( !dispatchUpdate ) {
-                const res: any = await getMenuData();
-                setDispatchUpdate(true);
-                dispatch(res);
-            }
-        };
-        fetchStoreMenu();
-        
-    }, [dispatchUpdate, dispatch]); 
 
     return (
         <>
@@ -70,7 +40,6 @@ const NestedRoutes = () => {
 
             <Layout
                 pageTitle="Nested Routes"
-                nav={JSON.stringify(storeData.menuItems)}
                 contentComponent={<><MainContent /></>}
             />