+++ pages/nested-routes/[id].tsx
@@ -1,13 +1,8 @@
 import Head from 'next/head';
 import { useRouter } from 'next/router';
-import { useEffect, useState } from 'react';
 import Link from 'next/link';
 import Layout from '@/components/Layout';
 
-
-// store
-import { useDispatch, useSelector } from "react-redux";
-import getMenuData from "@/store/actions/demoMenuActions";
 
 
 const MainContent = (props) => {
@@ -41,28 +36,6 @@ const NestedRoutesChild = () => {
     const id = router.query.id as string;
     const idShow = id.replace('.html', '')
 
-    // Get store
-    const [dispatchUpdate, setDispatchUpdate] = useState<boolean>(false);
-    const dispatch = useDispatch();
-    const storeData = useSelector((state: any) => {
-        return state.menuData;
-    });
-    
-    useEffect(() => {
-
-        // Get store
-        //-----
-        const fetchStoreMenu = async () => {
-            if ( !dispatchUpdate ) {
-                const res: any = await getMenuData();
-                setDispatchUpdate(true);
-                dispatch(res);
-            }
-        };
-        fetchStoreMenu();
-        
-    }, [dispatchUpdate, dispatch]); 
-
     return (
         <>
             <Head>
@@ -71,7 +44,6 @@ const NestedRoutesChild = () => {
 
             <Layout
                 pageTitle={idShow}
-                nav={JSON.stringify(storeData.menuItems)}
                 contentComponent={<><MainContent data={idShow} /></>}
             />