+++ pages/nested-routes/[id]/[comment].tsx
@@ -1,12 +1,6 @@
 import Head from 'next/head';
 import { useRouter } from 'next/router';
-import { useEffect, useState } from 'react';
 import Layout from '@/components/Layout';
-
-
-// store
-import { useDispatch, useSelector } from "react-redux";
-import getMenuData from "@/store/actions/demoMenuActions";
 
 
 const MainContent = (props) => {
@@ -32,30 +26,6 @@ const Comment = () => {
   
     const idShow = id.replace('.html', '')
 
-    // Get store
-    const [dispatchUpdate, setDispatchUpdate] = useState<boolean>(false);
-    const dispatch = useDispatch();
-    const storeData = useSelector((state: any) => {
-        return state.menuData;
-    });
-    
-
-
-    useEffect(() => {
-
-        // Get store
-        //-----
-        const fetchStoreMenu = async () => {
-            if ( !dispatchUpdate ) {
-                const res: any = await getMenuData();
-                setDispatchUpdate(true);
-                dispatch(res);
-            }
-        };
-
-        fetchStoreMenu();
-        
-    }, [dispatchUpdate, dispatch]); 
 
     return (
         <>
@@ -65,7 +35,6 @@ const Comment = () => {
 
             <Layout
                 pageTitle={idShow}
-                nav={JSON.stringify(storeData.menuItems)}
                 contentComponent={<><MainContent data={{idShow, comment}} /></>}
             />