+++ pages/posts.html.tsx
@@ -1,6 +1,5 @@
 import Head from 'next/head';
 import Link from 'next/link';
-import { useEffect, useState } from 'react';
 import Layout from '@/components/Layout';
 
 
@@ -8,27 +7,22 @@ import apiUrls from '@/config/apiUrls';
 import axios from "axios";
 
 
-// store
-import { useDispatch, useSelector } from "react-redux";
-import getMenuData from "@/store/actions/demoMenuActions";
-
-
 const MainContent = (props) => {
 
     return (
         <>
-              <ul>
+            <ul>
 
                 {!props.data ? <>Loading...</> : props.data.map((post: any) => {
-                  return (
-                    <li key={post.name}>
-                      <Link href={`/posts/${post.name}.html`} dangerouslySetInnerHTML={{__html: `${post.name} - (region: ${post.region})` }}></Link>
+                    return (
+                        <li key={post.name}>
+                            <Link href={`/posts/${post.name}.html`} dangerouslySetInnerHTML={{ __html: `${post.name} - (region: ${post.region})` }}></Link>
 
-                      
-                    </li>
-                  );
+
+                        </li>
+                    );
                 })}
-              </ul>
+            </ul>
 
         </>
     )
@@ -43,30 +37,6 @@ function Posts({ currentData }) {
 
     const posts = currentData;
 
-    // Get store
-    const [dispatchUpdate, setDispatchUpdate] = useState<boolean>(false);
-    const dispatch = useDispatch();
-    const storeData = useSelector((state: any) => {
-        return state.menuData;
-    });
-
-
-    useEffect(() => {
-
-        // Get store
-        //-----
-        const fetchStoreMenu = async () => {
-            if ( !dispatchUpdate ) {
-                const res: any = await getMenuData();
-                setDispatchUpdate(true);
-                dispatch(res);
-            }
-        };
-
-        fetchStoreMenu();
-        
-    }, [dispatchUpdate, dispatch]); 
-
     return (
         <>
             <Head>
@@ -76,8 +46,7 @@ function Posts({ currentData }) {
 
             <Layout
                 pageTitle="Posts"
-                nav={JSON.stringify(storeData.menuItems)}
-                contentComponent={<><MainContent data={posts}/></>}
+                contentComponent={<><MainContent data={posts} /></>}
             />
 
 
@@ -90,19 +59,18 @@ function Posts({ currentData }) {
  * ---------------------------------
 */
 export async function getStaticProps() {
-  
+
     const res: any = await axios.get(apiUrls.RECEIVE_DEMO_LIST);
-  
+
     // Pass data to the page via props
     return {
-      props: { 
-          currentData: res.data 
-      },
-  
-      // Incremental Static Regeneration. (Next.js will attempt to re-generate the page:)
-      revalidate: 10, // In seconds 
+        props: {
+            currentData: res.data
+        },
+
+        // Incremental Static Regeneration. (Next.js will attempt to re-generate the page:)
+        revalidate: 10, // In seconds 
     }
-  }
-  
-  export default Posts;
-  
+}
+
+export default Posts;