+++ src/components/BackToTop/index.tsx
@@ -158,7 +158,7 @@ export default function BackToTop(props: BackToTopProps) {
         }
 
 
-    }, []);
+    }, []); 
 
     return (
         <>