+++ src/components/Dashboard/Userinfo.tsx
@@ -11,7 +11,7 @@ const Userinfo = () => {
             setUsername(response);
         });
 
-    }, []); // Empty array ensures that effect is only run on mount and unmount
+    }, []); 
 
     return (
         <>